# Adds cyclestate.Pod to share unmarshalled result to all plugins