--- a/guest/internal/cyclestate/cyclestate.go
+++ b/guest/internal/cyclestate/cyclestate.go
@@ -1,0 +1,18 @@
+// Package cyclestate holds fields scoped to the Pod Scheduling context, e.g.
+// the beginning of the scheduling cycle until the end of the binding cycle.
+//
+// This field is reset on prefilter.Plugin, as that begins a new scheduling
+// cycle. Even if the pod is the same as a failed cycle, its state must be
+// reset so that any change is visible to the guest plugin.
+//
+// See https://kubernetes.io/docs/concepts/scheduling-eviction/scheduling-framework/#extension-points
+package cyclestate
+
+import (
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/prefilter"
+)
+
+// Pod is the current pod being scheduled. It is lazy and the same values are
+// returned for any plugins in a scheduling cycle.
+var Pod api.Pod = prefilter.Pod