… examples/filter-simple/main.wasm
… examples/prefilter-simple/main.wasm
… examples/score-simple/main.wasm
… guest/internal/prefilter/cstring.go
… guest/internal/prefilter/cstring_test.go
… guest/internal/prefilter/imports.go
… guest/internal/prefilter/imports_stub.go
… guest/internal/types/types.go
… guest/testdata/all-noop/main.wasm
… guest/testdata/all/main.wasm
… guest/testdata/filter/main.wasm
… guest/testdata/prefilter/main.wasm
… guest/testdata/score/main.wasm