@codefromthecrypt 2023-06-22
I think there's a bug in the test score-simple plugin setup as it is too fast. I'll look into it.

```
PluginScore/test/params:_real-12                          39818.0n ± 17%   350.1n ±  1%  -99.12% (p=0.002 n=6)
```