@codefromthecrypt 2023-06-22
ah right I didn't change the benchmarks to begin with PreFilter, so they were caching the result. fixing now