@codefromthecrypt 2023-06-23
// Intentionally change the pointer and run PreFilter, 
> 
> we don't need to change the pointer anymore right?
> #42 <https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/pull/42>
> 
> 
> —
> Reply to this email directly, view it on GitHub <https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/pull/45#pullrequestreview-1494207569>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAAPVV263OBMFEO3SHBPN33XMTXRDANCNFSM6AAAAAAZPXUWY4>.
> You are receiving this because you authored the thread.Message ID: ***@***.***>
> 


so this pointer isnt visible guest side. this is to make sure host side is like real life but agree we arent taking action so can remove